backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Production WSGI Server
gunicorn

# Performance Enhancements
uvloop
httptools

# Database
psycopg2-binary
asyncpg
sqlalchemy[asyncio]

# Authentication & Security
clerk-backend-api
python-jose[cryptography]
passlib[bcrypt]
svix

# Configuration & Environment
pydantic-settings
pydantic[email]
python-dotenv
python-multipart

# Caching
cachetools

# Testing (Development/CI only)
pytest
pytest-asyncio
